refactor: Removes Api::V1 namespace to meet the Instructions
